requirements.txt
pytest==9.0.2
requests==2.32.5
selene==2.0.0rc9
allure-pytest==2.15.2
webdriver-manager==4.0.2
allure-python-commons==2.15.2
selenium~=4.29.0